requirements.txt
anyio==4.0.0
astroid @ file:///croot/astroid_1676904296642/work
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama @ file:///croot/colorama_1672386526460/work
contourpy==1.1.1
cycler==0.12.1
dill @ file:///croot/dill_1692271232022/work
docstring-to-markdown @ file:///croot/docstring-to-markdown_1673447639232/work
exceptiongroup==1.1.3
filelock==3.12.4
flake8 @ file:///croot/flake8_1674581792275/work
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.43.1
fsspec==2023.9.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.5
isort @ file:///tmp/build/80754af9/isort_1628603791788/work
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
kiwisolver==1.4.5
lazy-object-proxy @ file:///home/builder/ci_310/lazy-object-proxy_1640791307593/work
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.1
mccabe @ file:///opt/conda/conda-bld/mccabe_1644221741721/work
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.25.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pdf2image==1.16.3
Pillow==10.0.1
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
pyclipper==1.3.0.post5
pycodestyle @ file:///croot/pycodestyle_1674267221883/work
pydocstyle @ file:///croot/pydocstyle_1675221668445/work
pyflakes @ file:///croot/pyflakes_1674165128613/work
pylint @ file:///croot/pylint_1676919896363/work
pyparsing==3.1.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server @ file:///croot/python-lsp-server_1681930392028/work
python-telegram-bot==20.5
pytoolconfig @ file:///croot/pytoolconfig_1676315026871/work
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rope @ file:///croot/rope_1676675015455/work
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
simpy==4.0.2
six==1.16.0
sniffio==1.3.0
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1637937080595/work
sympy==1.12
tifffile==2023.9.26
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tomlkit @ file:///tmp/abs_56_0lnnq5x/croots/recipe/tomlkit_1658946880479/work
torch==2.1.0
torchvision==0.16.0
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
ujson @ file:///opt/conda/conda-bld/ujson_1657544923770/work
urllib3==2.0.6
Werkzeug==3.0.1
whatthepatch @ file:///opt/conda/conda-bld/whatthepatch_1661795988879/work
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
xlrd==2.0.1
yapf @ file:///tmp/build/80754af9/yapf_1615749224965/work
